pyproject.toml
[project]
name = "alpha-verification-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.5",
    "aiosqlite>=0.22.1",
    "discord-py>=2.7.1",
    "httpx>=0.28.1",
    "phonenumbers>=9.0.27",
    "pillow>=12.2.0",
    "pymupdf>=1.27.2.2",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
]
